requirements.txt
Django~=3.2.9
djangorestframework~=3.13.1